+++ src/parsing/check_map_validity.c
@@ -11,8 +11,8 @@ void	get_angle(t_data *data, char c, int y, int x)
 		data->player->angle = M_PI;
 	else
 		data->player->angle = 0;
-	data->player->px = (float)x * TILE + TILE / 2;
-	data->player->py = (float)y * TILE + TILE / 2;
+	data->player->px = x * TILE + TILE / 2;
+	data->player->py = y * TILE + TILE / 2;
 }
 
 static int	check_char(t_data *data, char **map, int *y, int *x)
